requirements.txt
# Discord et communication
discord.py==2.5.2
aiohttp==3.11.14
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
PyNaCl>=1.5.0

# Configuration et environnement
python-dotenv>=1.0.0
colorama>=0.4.6
coloredlogs>=15.0.1
attrs==25.3.0

# YouTube et média
yt-dlp==2025.2.19
ffmpeg-python>=0.2.0
PyAudio>=0.2.14

# Graphiques et traitement de données 
matplotlib>=3.10.1
numpy>=2.0.2
pillow>=11.0.0
qrcode>=8.0.0

# Gestion du temps et formats
tzdata>=2024.1
python-dateutil>=2.9.0

# Outils de développement
black>=25.1.0
pylint>=3.3.6

# Gestion des requêtes et réseau
requests==2.32.3
urllib3==2.3.0
httpx>=0.28.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
multidict==6.2.0
yarl==1.18.3
propcache==0.3.0
frozenlist==1.5.0

# Utilitaires système
psutil>=7.0.0
platformdirs>=4.3.7
mcstatus>=10.0.0

# Recherche Wikipédia
wikipedia==1.4.0

# Autres dépendances
pytz
